Refactor: Reorganise some types
Move Cell types into Cell and the ClickHandler type to its own type
definition file.

.d.ts is the extension for a typescript-only file

Now all type imports point upwards, or to a shared type definition file.
