+++ src/clickHandler.d.ts
@@ -1,0 +1,3 @@
+import type { MouseEvent } from "react";
+
+export type ClickHandler<T> = (event: MouseEvent<T>) => void;
