+++ src/components/Cell.tsx
@@ -1,7 +1,11 @@
 import React from "react";
 import "./Cell.css";
 import classNames from "classnames";
-import { CellProps } from "../App";
+import type { ClickHandler } from "../clickHandler";
+
+export type CellClickHandler = ClickHandler<HTMLTableDataCellElement>;
+export type CellData = { word: string; stamped: boolean };
+export type CellProps = CellData & { toggleStamped: CellClickHandler };
 
 const Cell: React.FC<CellProps> = ({ word, stamped, toggleStamped }) => {
   const classes = classNames("Cell", { stamped: stamped });
