+++ src/lib/JsonSession.test.ts
@@ -1,5 +1,5 @@
 import JsonSession from "./JsonSession";
-import type { CellData } from "../App";
+import type { CellData } from "../components/Cell";
 
 beforeEach(() => {
   window.localStorage.clear();
